refactor: typescript 파일 분리
- 생산성, 가독성 향상을 위해 index.data.ts 파일을 분리 했습니다.

by @hyoloui
